+++ app/controllers/test/supplementary.controller.js
@@ -5,8 +5,8 @@
  * @module SupplementaryController
  */
 
-const FindRegionService = require('../../services/supplementary_billing/find_region.service.js')
-const SupplementaryService = require('../../services/supplementary_billing/supplementary.service.js')
+const FindRegionService = require('../../services/supplementary-billing/find-region.service.js')
+const SupplementaryService = require('../../services/supplementary-billing/supplementary.service.js')
 
 class SupplementaryController {
   static async index (request, h) {
