+++ app/lib/base-notifier.lib.js
