+++ app/lib/request-notifier.lib.js
@@ -5,7 +5,7 @@
  * @module RequestNotifierLib
  */
 
-const BaseNotifierLib = require('./base_notifier.lib.js')
+const BaseNotifierLib = require('./base-notifier.lib.js')
 
 /**
  * A combined logging and Airbrake (Errbit) notification manager for actions that take place within a
