+++ app/services/plugins/hapi-pino-ignore-request.service.js
