+++ app/services/supplementary-billing/supplementary.service.js
@@ -5,8 +5,8 @@
  * @module SupplementaryService
  */
 
-const BillingPeriodService = require('./billing_period.service.js')
-const FetchChargeVersionsService = require('./fetch_charge_versions.service.js')
+const BillingPeriodService = require('./billing-period.service.js')
+const FetchChargeVersionsService = require('./fetch-charge-versions.service.js')
 const SupplementaryPresenter = require('../../presenters/supplementary.presenter.js')
 
 class SupplementaryService {
