+++ test/lib/request-notifier.lib.test.js
@@ -9,7 +9,7 @@ const { describe, it, beforeEach, afterEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Thing under test
-const RequestNotifierLib = require('../../app/lib/request_notifier.lib.js')
+const RequestNotifierLib = require('../../app/lib/request-notifier.lib.js')
 
 describe('RequestNotifierLib class', () => {
   const id = '1234567890'
