+++ test/services/database-health-check.service.test.js
@@ -8,7 +8,7 @@ const { describe, it } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Thing under test
-const DatabaseHealthCheckService = require('../../app/services/database_health_check.service.js')
+const DatabaseHealthCheckService = require('../../app/services/database-health-check.service.js')
 
 describe('Database Health Check service', () => {
   it('confirms connection to the db by not throwing an error', async () => {
