+++ test/services/plugins/hapi-pino-ignore-request.service.test.js
@@ -12,7 +12,7 @@ const { expect } = Code
 const LogConfig = require('../../../config/log.config.js')
 
 // Thing under test
-const HapiPinoIgnoreRequestService = require('../../../app/services/plugins/hapi_pino_ignore_request.service.js')
+const HapiPinoIgnoreRequestService = require('../../../app/services/plugins/hapi-pino-ignore-request.service.js')
 
 describe('Hapi Pino Ignore Request service', () => {
   const _options = {}
